requirements.txt
# === 1. PyTorch (Install first) ===
torch==2.1.0
torchvision==0.16.0

# === 2. Python Essentials ===
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-multipart==0.0.6
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
joblib==1.3.2
Pillow==10.1.0
opencv-python==4.8.1.78
seaborn==0.13.2
xgboost==2.0.3

# === 3. OpenMMLab Dependencies ===
openmim
mmengine
# Use mim to install the following after mmengine is installed:
# mim install mmcv==2.1.0
# mim install mmdet==3.2.0
# mim install mmpose==1.2.0 --no-deps

# === 4. Additional Libraries ===
google-genai